backend/requirements.txt
Django==1.6.5
Pillow==2.5.1
djangorestframework==2.3.13
ipython==2.1.0
wsgiref==0.1.2
